pyproject.toml
[project]
name = "master-thesis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.8.0",
    "jiwer>=3.1.0",
    "marimo>=0.11.30",
    "matplotlib>=3.10.1",
    "mistralai>=1.6.0",
    "openai>=1.69.0",
    "pillow>=11.1.0",
    "pre-commit>=4.2.0",
    "pydantic>=2.11.0",
    "python-dotenv>=1.1.0",
    "seaborn>=0.13.2",
]
